delete dtype

+++ build.py
@@ -69,7 +69,7 @@ def nn_compile(name, embed_mat, seq_len, class_num):
     vocab_num, embed_len = embed_mat.shape
     embed = Embedding(input_dim=vocab_num, output_dim=embed_len,
                       weights=[embed_mat], input_length=seq_len, trainable=True)
-    input = Input(shape=(seq_len,), dtype='int32')
+    input = Input(shape=(seq_len,))
     embed_input = embed(input)
     func = map_item(name, funcs)
     if name == 'rnn_bi_crf':
